Rewrote Player Controller
The PlayerController class has been entirely rewritten.  All properties and functions are capitalized and stylized correctly.  All properties/variables have been privatized/encapsulated.  All necessary functions have also been privatized.  Changed all configurable properties to [SerializedField]s and added Headers.  Added property for IsGrounded() checkSphere radius.  Also added property for IsGrounded() checkSphere Y offset, to adjust the centerpoint of the checkSphere, which is normally centered off of the bounds.extents.y of the player capsule collider (position of the feet).  The offset value is added to this: a positive float moves it up, a negative float moves it down.  The use of AudioClips in the two state-based movement functions is now optional, through the use of conditional statements that check if they have been assigned or not.  This controller uses the new Look(), Move(), Dash(), Jump(), and IsGrounded() Functions.  Closes #24.