Added RoomCleanup() to RoomDaemon
The RoomDaemon now has a RoomCleanup() function that relies on roomIsClear and the distance between the player and the exit of the room, taken in as a Serialized Property.  Old rooms will be cleaned up using Destroy(), but only if both the room has been cleared, and the player has traveled a certain distance away.  This should keep memory usage low, by reducing the number of meshes, rigidbodies, and other components in the scene.  Each cleanupDistance property is set uniquely on each room prefab.  Closes #35.