Fixed UI & Removed RoomCleanup(), PGS Now Cleans up Old Rooms
I removed all the code added to the RoomDaemon on the last commit, instead, I kept the line of code that I added to the ProceduralGenerationSystem class.  It set the parent of the last hallwayInstance to the currentRoom.  Then I added a new line of code to generateNextRoom(), Destroy(lastRoom) before lastRoom is redefined.  This only keeps two rooms in the scene at a time.

Also fixed UI scaling, now scales with resoltuion. Fixes #39