Added BossBar Prefab and Fixed a Few Bugs
Added a BossBar, an easing health bar for bosses.  It is in the UIContainer prefab.  It has a script that disables/enables it's child object when a boss dies/is spawned in.

I fixed a bug that would not allow the player's cursor to unlock when the player died.  This would prevent the player from being able to click the restart button.

Started working on #42 by testing on the pistol.  Still a work in progress