Enemies and Bosses now Drop loot
Fixed a bug where the FPS meter would miscalculate the current FPS when the game was paused.  It now checks to see if the timescale is 0 before calculating and displaying the FPS.

Enemy and Boss Prefabs now have the DropLoot component.  It is assigned a pickupItem prefab and a drop rate in the form of an int as a percentage.  Right now Enemies have a 10% chance to drop health.  The VSauceBoss has a 100% chance to drop the railGun.

The railGun pickupItem prefab is no longer in the spawnRoom in the MainScene.