Added EXProver
Added new enemy EXProver.  It runs towards the player and detonates.  A delay needs to be added to the attack, and the light object on the head needs to be integrated into the attack animation.  Everything else about this enemy is fully implemented.  The ragdoll, AI, healthcontroller, and barrel all work. Closes #41